Refactored to valid Java naming convention. Changed to valid utility convention. Renamed to proper classname in test/src. Parameters moved to resources
