+++ src/main/java/manager/DriverManager.java
@@ -1,10 +1,10 @@
-package DriverManager;
+package manager;
 
 import org.openqa.selenium.WebDriver;
 
 public abstract class DriverManager {
 
-    protected WebDriver driver;
+    WebDriver driver;
     protected abstract void startService();
     protected abstract void stopService();
     protected abstract void createDriver();
