+++ src/main/java/manager/DriverManagerFactory.java
@@ -1,9 +1,11 @@
-package DriverManager;
+package manager;
 
 import java.lang.reflect.InvocationTargetException;
 
-public class DriverManagerFactory {
+public final class DriverManagerFactory {
 
+    private DriverManagerFactory() {
+    }
 
     public static <T extends DriverManager> T getManager(final Class<T> type) throws NoSuchMethodException, IllegalAccessException, InvocationTargetException, InstantiationException {
         return type.getConstructor().newInstance();
