+++ src/main/java/manager/FirefoxDriverManager.java
@@ -1,4 +1,4 @@
-package DriverManager;
+package manager;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.firefox.FirefoxDriver;
