+++ src/main/java/pages/BasePage.java
@@ -1,4 +1,4 @@
-package Pages;
+package pages;
 
 import org.openqa.selenium.WebDriver;
 
@@ -7,16 +7,16 @@ abstract class BasePage<T extends BasePage> {
     private WebDriver webDriver;
 
     protected abstract String getPageURL();
-    protected abstract T getPage();
+    protected abstract T getThis();
 
     public T setWebDriver(WebDriver webDriver) {
         this.webDriver = webDriver;
-        return getPage();
+        return getThis();
     }
 
     public T openPage() {
         webDriver.get(getPageURL());
-        return getPage();
+        return getThis();
     }
 
     public final String getPageTitle() {
